Portco Revenue Chart (#129)
This is part one of a number of PRs for the revenue chart and API integration. In this PR the API is called and saved in a Redux store. Selectors are used to grab the data and populate the sections 1.1, 1.2 and 1.3 of the corresponding SS: https://casertaconcepts.atlassian.net/wiki/spaces/SOF/pages/522945521/PortCo+Dashboard+Revenue+Widget+-+SS